requirements.txt
openai
python-dotenv
pygments
